refactor: extract time-picker formatTime and parseTime helpers
